Add Carousel functionality to Review component
